Resources_RestAPI/requirements.txt
fastapi
uvicorn
motor
sqlalchemy
python-dotenv
pydantic
langchain_huggingface
langchain-community
mysql-connector-python